Query Data Chart updates and fixes
Query Data Chart now adds in line markers and properly handles all types
of query chart series, as well as displaying a dotted line to show gaps
in the graph. Also, code has been heavily refactored -- could use a
little more work on that though, but it is much more manageable. Lastly,
chart labels are broken and need to be transitioned to the new "method"
of doing them as I had come up with for graphing points -- will be done
ASAP. Data tags for MED will need to be reimplemented as well for
patient snapshot.